backend/requirements.txt
astroid==2.2.5
autopep8==1.4.4
Click==7.0
Flask==1.0.2
Flask-Jsonpify==1.5.0
isort==4.3.20
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2-binary==2.8.2
pycodestyle==2.5.0
pylint==2.3.1
six==1.12.0
typed-ast==1.3.5
Werkzeug==0.15.4
wrapt==1.11.1
